Add i3, reorganize makefile
-- i3 --
This adds my normal i3 config along with scripts for jumping to
workspaces with <num>: prefixes.

-- Makefile --
The Makefile has been reorganized to make it a little easier to add
targets. This would still be better off replaced with stow(1)
eventually, but for now this works fine.

The default target is now the help target, which will print help text
for each target and any variables that are known about the system.